+++ src/lib/immer.js
@@ -1,0 +1,9 @@
+// immer uses a non standard global property from Node.js "process"
+window.process = {
+  env: {
+    NODE_ENV: 'production',
+  },
+};
+
+export { produce } from 'immer/dist/immer.esm.js';
+// TODO: object freeze